modularized TicketList and Ticket components, added employeeService module and getAllEmployees fetch function, edited getAllTickets fetch call to embed employeeTickets to extract employeeId foreign key to get employee name, modularized CSS for Tickets directory, imported Ticket components to App.jsx
