+++ src/services/employeeService.js
@@ -1,0 +1,4 @@
+export const getAllEmployees = async () => {
+	const response = await fetch(`http://localhost:8088/employees?_expand=user`)
+	return response.json()
+}
