+++ src/services/ticketService.js
@@ -1,4 +1,6 @@
 export const getAllTickets = async () => {
-	const response = await fetch(`http://localhost:8088/serviceTickets`)
+	const response = await fetch(
+		`http://localhost:8088/serviceTickets?_embed=employeeTickets`
+	)
 	return response.json()
 }
